Moved assertions under PageObjects & removed Specifications folders
