+++ src/TestStack.Seleno.AcceptanceTests/PageObjects/Actions/ModelIoTests.cs
@@ -3,7 +3,7 @@ using TestStack.BDDfy;
 using TestStack.Seleno.AcceptanceTests.Web.Fixtures;
 using TestStack.Seleno.AcceptanceTests.Web.ViewModels;
 
-namespace TestStack.Seleno.AcceptanceTests.PageObjects.ActionsTests
+namespace TestStack.Seleno.AcceptanceTests.PageObjects.Actions
 {
     abstract class ModelIoTests
     {
