+++ src/TestStack.Seleno/Configuration/Registration/RegistrationExtensions.cs
@@ -4,7 +4,7 @@ using OpenQA.Selenium;
 using TestStack.Seleno.Configuration.Contracts;
 using TestStack.Seleno.PageObjects;
 using TestStack.Seleno.PageObjects.Actions;
-using TestStack.Seleno.Specifications.Assertions;
+using TestStack.Seleno.PageObjects.Assertions;
 
 namespace TestStack.Seleno.Configuration.Registration
 {
