+++ src/TestStack.Seleno/PageObjects/UiComponent.cs
@@ -2,8 +2,8 @@
 using OpenQA.Selenium;
 using TestStack.Seleno.Configuration.Contracts;
 using TestStack.Seleno.PageObjects.Actions;
+using TestStack.Seleno.PageObjects.Assertions;
 using TestStack.Seleno.PageObjects.Controls;
-using TestStack.Seleno.Specifications.Assertions;
 
 namespace TestStack.Seleno.PageObjects
 {
